pyproject.toml
[tool.poetry]
name = "superlogica-teste"
version = "0.1.0"
description = ""
authors = ["Ruda <filiperuda@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyspark = "^3.5.3"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
